pyproject.toml
[tool.poetry]
name = "playlist-export-sample"
version = "0.1.0"
description = ""
authors = ["manabreakmura <174273321+manabreakmura@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
